reference/scenarios/langchain/pyproject.toml
[project]
name = "langchain-reference-test"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.15",
    "langchain-openai==1.1.12",
    "openai==2.31.0",
    "genai-reference-shared",
]

[tool.uv.sources]
genai-reference-shared = { path = "../../shared", editable = true }

[tool.uv]
package = false
